EmployeeNotification và ....

+++ EVService/src/main/java/edu/uth/evservice/EVService/model/ServicePackage.java
@@ -14,7 +14,7 @@ public class ServicePackage {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long packageId;
+    private Integer packageId;
 
     @Column(nullable = false)
     private String packageName;
