+++ EVService/src/main/java/edu/uth/evservice/EVService/repositories/IServicePackageRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IServicePackageRepository extends JpaRepository<ServicePackage, Long> {
+public interface IServicePackageRepository extends JpaRepository<ServicePackage, Integer> {
 }
